// backend/src/feature-flag/enums/feature-flag.enums.ts
export enum RolloutStrategy {
  ALL = 'ALL',
  PERCENTAGE = 'PERCENTAGE',
  MERCHANT_IDS = 'MERCHANT_IDS',
  MERCHANT_TIERS = 'MERCHANT_TIERS',
  COUNTRIES = 'COUNTRIES',
}

export enum FeatureFlagEvaluationReason {
  DISABLED_GLOBALLY = 'DISABLED_GLOBALLY',
  OVERRIDE_ON = 'OVERRIDE_ON',
  OVERRIDE_OFF = 'OVERRIDE_OFF',
  ALL_ENABLED = 'ALL_ENABLED',
  PERCENTAGE_ROLLOUT = 'PERCENTAGE_ROLLOUT',
  PERCENTAGE_EXCLUDED = 'PERCENTAGE_EXCLUDED',
  MERCHANT_IDS_MATCH = 'MERCHANT_IDS_MATCH',
  MERCHANT_IDS_NO_MATCH = 'MERCHANT_IDS_NO_MATCH',
  MERCHANT_TIER_MATCH = 'MERCHANT_TIER_MATCH',
  MERCHANT_TIER_NO_MATCH = 'MERCHANT_TIER_NO_MATCH',
  COUNTRY_MATCH = 'COUNTRY_MATCH',
  COUNTRY_NO_MATCH = 'COUNTRY_NO_MATCH',
}